extract ErizoJSManager
